\name{calc.camp.coverage}
\alias{calc.camp.coverage}
\title{Calculate the expected coverage of each campaign given...}
\usage{calc.camp.coverage(camps, trans.rho, alpha)}
\description{Calculate the expected coverage of each campaign given
the trans.rho and alpha.}
\details{Corresponds to equation 5 in paper}
\value{TBD}
\arguments{\item{camps}{TBD}
\item{trans.rho}{TBD}
\item{alpha}{TBD}}
