\name{mcmc.converge}
\alias{mcmc.converge}
\title{Function to assess MCMC convergance of a chain given a matrix of MCMC...}
\usage{mcmc.converge(chains)}
\description{Function to assess MCMC convergance of a chain given a matrix of MCMC
iterations.}
\value{TBD}
\arguments{\item{chains}{a matrix of n.interations X n.chains with results}}
