\name{vacem-package}
\alias{vacem-package}
\title{Vaccination Activities Coverage Estimation Model}
\description{A likelihood framework for estimating the effective coverage of vaccination
programs using cross-sectional surveys combined with administrative data.}
\details{\tabular{ll}{ 
Package:  \tab vacem\cr 
Type:     \tab Package\cr 
Version:  \tab 0.1-1\cr 
Date:     \tab 2012-01-30\cr 
License:  \tab GPL (>= 2)\cr 
LazyLoad: \tab yes\cr 
} 

The performance of routine and supplemental immunization activities is
usually measured by the administrative method: dividing the number of doses
distributed by the size of the target population. This package implements
a method that combines the administrative data with measurements from
cross-sectional surveys of vaccine coverage, e.g. Demographic and Health
Surveys (DHS), to provide an improved estimate of vaccination coverage.}
\docType{package}
\author{Justin Lessler \email{jlessler@jhsph.edu}
Jessica Metcalf \email{cmetcalf@princeton.edu}

Maintainer: Ken Cline \email{kcline@jhsph.edu}}
\references{Lessler J, Metcalf CJE, Grais RF, Luquero FJ, Cummings DAT,
et al. (2011) Measuring the Performance of Vaccination Programs
Using Cross-Sectional Surveys: A Likelihood Framework and
Retrospective Analysis. PLoS Med 8(10): e1001110.
doi:10.1371/journal.pmed.1001110
\url{ http://www.plosmedicine.org/article/info:doi/10.1371/journal.pmed.1001110 }}
\keyword{vaccine}
\keyword{immunization}
\keyword{likelihood}
\keyword{models}
\examples{# TBD}
